requirements.txt
disnake
disnake[voice]
git+https://github.com/lanegramling/youtube-dl@fix-function-sig-parser#egg=youtube-dl
motor==2.5.1
pymongo
pymongo[srv]
requests
unidecode
python-decouple
gTTS
google-api-python-client
wikipedia
wikipedia-api
Pillow
deep-translator
colorthief
aiohttp
aiofiles
markovchain
markovify